Select a development model structure

The function calls a model table with the parameterizations for different species from different studies built in. Refer to the table (model_table) before using function to find inputs for the different function arguments. It pulls the model format as a string and parses it to be usable in hatchR model.

Usage

model_select(author, species, model_id, development_type = "hatch")

Arguments

author

Character string of author name.

species

Character string of species name.

model_id

Either model number from Beacham and Murray (1990) or specific to other paper (e.g., Sparks et al. 2017 = AK).

development_type

The phenology type. A vector with possible values "hatch" or "emerge". The default is "hatch".

Value

A data.frame giving model specifications to be passed to predict_phenology().

Examples

library(hatchR)
# access the parameterization for sockeye hatching using
# model #2 from Beacham and Murray (1990)
sockeye_hatch_mod <- model_select(
  author = "Beacham and Murray 1990",
  species = "sockeye",
  model_id = 2,
  development_type = "hatch"
)
# print
sockeye_hatch_mod
#> # A tibble: 1 × 5
#>   author                  species model_id development_type expression          
#>   <chr>                   <chr>   <chr>    <chr>            <chr>               
#> 1 Beacham and Murray 1990 sockeye 2        hatch            1 /exp(6.727 - log(…