Table of phenology models

Description to come...

Usage

model_table

Format

Data frame

author

Short article citation

species

Species name

model

Model number from B&M

dev.type

Hatch or emergence

func

character string desciption of the function

Source

To come...

References

Beacham, T.D., Murray, C.B. (1990). Temperature, egg size, and development of embryos and alevins of five species of Pacific salmon: a comparative analysis. Canadian Journal of Zoology, 68, 1931–1940.

Sparks, M.M., Falke, J.A., Quinn, T.A., Adkinson, M.D., Schindler, D.E. (2019). Influences of spawning timing, water temperature, and climatic warming on early life history phenology in western Alaska sockeye salmon. Canadian Journal of Fisheries and Aquatic Sciences, 76(1), 123–135.

Austin, C.C., Essington, T.E., Quinn, T.A. (2019). Spawning and emergence phenology of bull trout Salvelinus confluentus under differing thermal regimes. Canadian Journal of Fisheries and Aquatic Sciences, 94(1), 191–195.

Examples

model_table
#> # A tibble: 51 × 5
#>    author                  species  model dev.type func                         
#>    <chr>                   <chr>    <chr> <chr>    <chr>                        
#>  1 Beacham and Murray 1990 sockeye  2     hatch    1 /exp(6.727 - log(x + 2.394…
#>  2 Beacham and Murray 1990 sockeye  4     hatch    1/exp(8.734 + (-1.589*log(x …
#>  3 Beacham and Murray 1990 sockeye  5     hatch    1/exp(9.848 + (-1.583*log(x …
#>  4 Beacham and Murray 1990 sockeye  7     hatch    1/exp(5.379 + (-0.076*log(x)…
#>  5 Beacham and Murray 1990 sockeye  2     emerge   1 / exp(7.227 - log(x + 2.56…
#>  6 Beacham and Murray 1990 sockeye  4     emerge   1/exp(7.647 + (-1.134*log(x …
#>  7 Beacham and Murray 1990 sockeye  5     emerge   1/exp(9.404 + (-1.169*log(x …
#>  8 Beacham and Murray 1990 sockeye  7     emerge   1/exp(5.97 + (-0.276*log(x))…
#>  9 Beacham and Murray 1990 pink odd 2     hatch    1 / exp(6.663 - log(x + 2.55…
#> 10 Beacham and Murray 1990 pink odd 4     hatch    1/exp(7.962 + (-1.382*log(x …
#> # ℹ 41 more rows