Example dataset: Lake Iliamna, Alaska

Temperature Data from Lake Iliamna on Woody Island, Alaska

Usage

woody_island

Format

Data frame with columns:

date

Date measurement was taken.

temp_c

Temperature in degrees Celsius.

Source

Sparks et al. (2019)

References

Sparks, M.M., Falke, J.A., Quinn, T.A., Adkinson, M.D., Schindler, D.E. (2019). Influences of spawning timing, water temperature, and climatic warming on early life history phenology in western Alaska sockeye salmon. Canadian Journal of Fisheries and Aquatic Sciences, 76(1), 123–135.

Examples

woody_island
#> # A tibble: 735 × 2
#>    date       temp_c
#>    <date>      <dbl>
#>  1 1990-08-11  25.8 
#>  2 1990-08-12  23.3 
#>  3 1990-08-13  18.5 
#>  4 1990-08-14  15.4 
#>  5 1990-08-15  14.0 
#>  6 1990-08-16  11.4 
#>  7 1990-08-17   9.19
#>  8 1990-08-18   9.84
#>  9 1990-08-19  10.6 
#> 10 1990-08-20  10.7 
#> # ℹ 725 more rows