Table of phenology models

Table providing the model parameterizations for the phenology models.

Usage

model_table

Format

Data frame with 5 columns:

author

Author-year short citation

species

Species name

model_id

Identification number for a model parameterization

development_type

Hatch or emergence

expression

character string of parameterized function expression

Source

Beacham & Murray (1990), Sparks et al. (2017), Austin et al. (2019).

References

Beacham, T.D., Murray, C.B. (1990). Temperature, egg size, and development of embryos and alevins of five species of Pacific salmon: a comparative analysis. Canadian Journal of Zoology, 68, 1931–1940.

Sparks, M.M., Westley, P.A.H., Falke, J.A., Quinn, T.A. (2017). Thermal adaptation and phenotypic plasticity in a warming world: Insights from common garden experiments on Alaskan sockeye salmon. Canadian Journal of Fisheries and Aquatic Sciences, 76(1), 123–135.

Austin, C.C., Essington, T.E., Quinn, T.A. (2019). Spawning and emergence phenology of bull trout Salvelinus confluentus under differing thermal regimes. Canadian Journal of Fisheries and Aquatic Sciences, 94(1), 191–195.

Examples

model_table
#> # A tibble: 51 × 5
#>    author                  species  model_id development_type expression        
#>    <chr>                   <chr>    <chr>    <chr>            <chr>             
#>  1 Beacham and Murray 1990 sockeye  2        hatch            1 /exp(6.727 - lo…
#>  2 Beacham and Murray 1990 sockeye  4        hatch            1/exp(8.734 + (-1…
#>  3 Beacham and Murray 1990 sockeye  5        hatch            1/exp(9.848 + (-1…
#>  4 Beacham and Murray 1990 sockeye  7        hatch            1/exp(5.379 + (-0…
#>  5 Beacham and Murray 1990 sockeye  2        emerge           1 / exp(7.227 - l…
#>  6 Beacham and Murray 1990 sockeye  4        emerge           1/exp(7.647 + (-1…
#>  7 Beacham and Murray 1990 sockeye  5        emerge           1/exp(9.404 + (-1…
#>  8 Beacham and Murray 1990 sockeye  7        emerge           1/exp(5.97 + (-0.…
#>  9 Beacham and Murray 1990 pink odd 2        hatch            1 / exp(6.663 - l…
#> 10 Beacham and Murray 1990 pink odd 4        hatch            1/exp(7.962 + (-1…
#> # ℹ 41 more rows